requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
altair
bcrypt

# BigQuery (read-only)
google-cloud-bigquery
google-auth
pyarrow
